% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_allocation.R
\name{calculate_allocation}
\alias{calculate_allocation}
\title{Capital Allocation}
\usage{
calculate_allocation(stanfit, ...)
}
\arguments{
\item{stanfit}{a \code{stanfit} object of outputs from \code{stan_bsrmmlogt}
or \code{stan_bsrmmlogn}}

\item{p}{percentile for expected shortfall}
}
\value{
a list with two elements:\code{risk measures} capital charges
for each segment and \code{allocation} percentage of capital allocation
}
\description{
Calculate capital allocation using Euler Principle
based on expected shortfall measures
}
