% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateOpLosses.R
\name{rLogGtLb}
\alias{rLogGtLb}
\title{Truncated Log Generalized t Random Variables}
\usage{
rLogGtLb(n, mu, sigma, nu, H = 0)
}
\arguments{
\item{n}{an integer giving the number of simulations}

\item{mu}{a numeric value giving the location parameter for the log
generalized t distribution}

\item{sigma}{a numeric value giving the shape parameter for the log
generalized t distribution}

\item{nu}{a numeric value giving the degree-of-freedom parameter for the log
generalized t distribution}

\item{H}{a numeric value giving the reporting threshold; default to 0.}
}
\value{
a sequence of numerical values of length \code{n} of random draws from
Log Generalized t (\code{mu}, \code{sigma}, \code{nu})
}
\description{
Simulate from left-truncated generalized log t distribution
}
\examples{
rLogGtLb(n = 10, mu = 10, sigma = 2, nu = 100, H = 0)
rLogGtLb(n = 100, mu = 12, sigma = 2, nu = 20, H = 5000)
}
