% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateOpLosses.R
\name{simulateOpLosses}
\alias{simulateOpLosses}
\title{Loss Distribution Approach}
\usage{
simulateOpLosses(nPeriods, lambda, mu, sigma, nu, H = 0, seed = 42)
}
\arguments{
\item{nPeriods}{An integer giving the number of periods to simulate
the losses}

\item{lambda}{An integer giving the average number of counts per time period}

\item{mu}{a numeric value giving the location parameter for the log
generalized t distribution}

\item{sigma}{a numeric value giving the shape parameter for the log
generalized t distribution}

\item{nu}{a numeric value giving the degree-of-freedom parameter for the log
generalized t distribution}

\item{H}{a numeric value giving the reporting threshold; default to 0.}
}
\value{
tibble data frame of loss data at event level.  The first column
indicates the time period; the second column indicates the loss amount
}
\description{
This function simulates operational losses following a commonly used loss
distribution approach.  Loss counts for each time period is simulated from
Poisson distribution with \code{lambda}.  Each loss amount per an loss event
is simulated from a log generalized t distribution.
}
\examples{
simulateOpLosses(nPeriods = 10, lambda = 5, mu = 10, sigma = 2, nu = 20, H = 5000)
}
