% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_bsrmmlogt.R
\name{stan_bsrmmlogt}
\alias{stan_bsrmmlogt}
\title{Bayesian Structural Risk Measurement Model - Log t likelihood}
\usage{
stan_bsrmmlogt(
  losses,
  sbu,
  time,
  bu = NULL,
  n_t = NULL,
  n_sbu = NULL,
  n_bu = NULL,
  frequency_prior = list(a_lambda, b_lambda, r_diric),
  severity_prior = list(a_nu = 3, b_nu = 0.1, scale_sigma = 10, scale_intercept = 10,
    m0 = length(unique(sbu)), sigma_tilde = NULL, slab_scale = 25, slab_df = 25),
  H = 1,
  pars = c("lambdas", "mu", "sigma", "nu", "Y_hat"),
  ...
)
}
\arguments{
\item{losses}{a sequence of losses}

\item{sbu}{a sequence of integers indicating which sbu generates a loss.  This
should be of the same length as \code{losses}}

\item{bu}{an optional sequence of the same length as n_sbu}

\item{n_t}{an optional integer allows for manually setting the number of time periods
in case the last time period has no loss in the data.}

\item{n_sbu}{an optional integer allows for manually setting number of sbu in case
the last sbu has no loss in the data.}

\item{n_bu}{an optional integer allows for manually setting number of bu in case
the last bu has no loss in the data.}

\item{frequency_prior}{a list of three prior parameters that characterize
frequency prior. \code{a_lambda} and \code{b_lambda} are unit-length numerics
determining the Gamma prior on the expected frequency at the enterprise level.
\code{r_diric} is a sequence of the same length as the number of sbu, determining
how the frequency at the enterprise level is distributed across sbu.}

\item{severity_prior}{a list of prior parameters for the prior distributions
on the parameters of the log generalized t likelihood.  \code{a_nu} and \code{b_nu}
are numerics of length 1.  \code{a_nu} and \code{b_nu} determines the Gamma
prior on the degree-of-freedom parameter \code{nu}.

\deqn{nu ~ G(a_nu, b_nu)}

Default to a weakly informative prior with \code{a_nu} = 3 and \code{b_nu} = 0.1.

\code{scale_sigma} is the prior parameter on the scale parameter of
half-Cauchy prior on the scale parameter of the log generalized t likelihood.

\deqn{sigma ~ halfCauchy(scale_sigma)}

Default to \code{scale_sigma} = 10}

\item{pars}{an optional string sequence in \code{rstan::sampling} to select
parameters to extract.  \code{lambdas}, \code{mu}, \code{sigma}, and \code{nu}
are required to construct the posterior distribution.}
}
\description{
Fit Bayesian Structural Risk Measurement Model using stan.
This function is a wrapper function around \emph{bstrmmlogt.stan}.
}
